refactor(inventory): rename sqlite modules
Module and struct names must follow the implementation technology, which
is sqlx
