+++ src/infra/mod.rs
@@ -1,1 +1,1 @@
-pub mod sqlite;
+pub mod sqlx;
